manpower-forecast/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
pandas==2.2.3
python-dotenv==1.0.1
alembic==1.14.0
